+++ lib/require_profiler/printer/text.rb
@@ -8,7 +8,7 @@ module RequireProfiler
       def flush(node, indent: 0)
         return unless flush?(node)
 
-        path = node.path.sub(prefix_stripper, "")
+        path = strip_prefix(node.path)
         output << "#{PAD * indent}#{path} — #{time_to_duration(node.time)}\n"
         node.children.each { flush(_1, indent: indent + 1) }